Fixed lighting with irradiance formula

+++ src/camera.h
@@ -11,15 +11,15 @@ public:
           c3(glm::dvec3(0.0f, 1.0f, 1.0f)),
           c4(glm::dvec3(0.0f, -1.0f, 1.0f)),
           width(width), height(height) {
-        this->pixelSize = 2.0f / height;
+        this->pixelSize = 2.0 / height;
     }
 
     Ray getRay(int i, int j) {
-        float x = i * pixelSize - (1.0 - pixelSize); // see lecture 4
-        float y = (height - j) * pixelSize - (1.0 - pixelSize); // had to go to -j to flip it 
+        double x = i * pixelSize - (1.0 - pixelSize); // see lecture 4
+        double y = (height - j) * pixelSize - (1.0 - pixelSize); // had to go to -j to flip it 
 
         glm::dvec3 pixelPosition = glm::dvec3(0.0, x, y);
-        glm::dvec3 rayDirection = pixelPosition - startPosition;
+        glm::dvec3 rayDirection = pixelPosition - startPosition; // random offset for anti aliasing
         return Ray(startPosition, rayDirection);
     }
 
@@ -27,6 +27,6 @@ public:
     glm::dvec3 direction;
     int width;
     int height;
-    float pixelSize;
+    double pixelSize;
     glm::dvec3 c1, c2, c3, c4;
 };
